backend/requirements.txt
Django
python-dotenv
django-seed
djangorestframework
mysqlclient